requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.1
pytesseract==0.3.10
openai==0.28.1
python-dotenv==1.0.0
gunicorn==21.2.0 